% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greg.R
\name{greg}
\alias{greg}
\title{Reguarlization graphical model estimation}
\usage{
greg(z, A, eps = 1e-15, rholist = NULL, gamma = 0.5, trace = FALSE)
}
\arguments{
\item{z}{n * p dimensional matrix}

\item{A}{p * p true graph}

\item{eps}{a tolerence level for thresholding}

\item{rholist}{a sequence of penalty parameters}

\item{gamma}{the adaptive lasso penalty parameter}

\item{trace}{whether to trace to estimation process.}
}
\value{
a list.
\item{roc.lasso}{roc results for lasso}
\item{roc.scad}{roc results for scad}
\item{roc.alasso}{roc results for adaptive lasso}
}
\description{
\code{greg} calculate regularization graphical model
}
\seealso{
\code{\link{pgraph}}, \code{\link{roc}}, \code{\link{projcov}}, \code{\link{projcore}}
}

