% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgraph.R
\name{pgraph}
\alias{pgraph}
\title{Compute the Projected Graph}
\usage{
pgraph(z, f = NULL, method = c("lasso", "sam"), cond = TRUE,
  trace = FALSE)
}
\arguments{
\item{z}{n * p dimensional matrix}

\item{f}{n * q factor matrix. Default = 'NULL'.}

\item{method}{projection method. Default = 'linear'.}

\item{cond}{whether to create a conditional graph or unconditional graph.
Default = TRUE. If cond = FALSE, f must be provided.}

\item{trace}{whether to trace to estimation process.}
}
\value{
a list.
\item{test}{distance covariance test object}
\item{xeps}{residual of the projection of x}
\item{yeps}{residual of the projection of y}
}
\description{
\code{pgraph} calculate the projected graph
}
\seealso{
\code{\link{greg}}, \code{\link{roc}}, \code{\link{projcov}}, \code{\link{projcore}}
}

