% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projcov.R
\name{projcov}
\alias{projcov}
\title{Calculate the Projected covariance of two vectors}
\usage{
projcov(x, y, b, method = c("lasso", "sam"))
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}

\item{b}{factor matrix}

\item{method}{projection method. Default = 'linear'.}
}
\value{
a list.
\item{test}{distance covariance test object}
\item{xeps}{residual of the projection of x}
\item{yeps}{residual of the projection of y}
}
\description{
\code{projcov} calculate the projected distance covariance of two vectors given
factors.
}
\seealso{
\code{\link{greg}}, \code{\link{roc}}, \code{\link{pgraph}}, \code{\link{projcore}}
}

