% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgraph.R
\name{pgraph}
\alias{pgraph}
\title{Compute the Projected Graph}
\usage{
pgraph(z, f = NULL, method = c("ridge", "lasso", "sam"), cond = TRUE,
  trace = FALSE)
}
\arguments{
\item{z}{n * p dimensional matrix}

\item{f}{n * q factor matrix. Default = 'NULL'.}

\item{method}{projection method. Default = 'linear'.}

\item{cond}{whether to create a conditional graph or unconditional graph.
Default = TRUE. If cond = FALSE, f must be provided.}

\item{trace}{whether to trace to estimation process.}
}
\value{
a list.
\item{test}{distance covariance test object}
\item{xeps}{residual of the projection of x}
\item{yeps}{residual of the projection of y}
}
\description{
\code{pgraph} calculate the projected graph
}
\examples{

p=30;
n=120;
tmp=runif(p-1,0.5,1)
s=c(0,cumsum(tmp));
s1=matrix(s,p,p)
cov.mat.true=exp(-abs(s1-t(s1)))
prec.mat.true=solve(cov.mat.true);
a=matrix(rnorm(p*n),n,p)
data.sa=a\%*\%chol(cov.mat.true);
fit = pgraph(data.sa)
fit2 = pgraph(data.sa, cor = 'dcov')
}

