% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projcore.R
\name{projcore}
\alias{projcore}
\title{Calculate the Projected matrix given factors}
\usage{
projcore(x, b, method = c("lasso", "sam"), one.SE = TRUE, refit = TRUE,
  randSeed = 0)
}
\arguments{
\item{x}{first vector}

\item{b}{factor matrix}

\item{method}{projection method. Default = 'linear'.}

\item{one.SE}{whether to use the 1se rule for glmnet. Default = TRUE.}

\item{refit}{whether to refit the selected model. Default = TRUE.}

\item{randSeed}{the random seed for the program. Default = 0.}
}
\value{
\item{eps}{the residual matrix after projection}
}
\description{
\code{projcore} calculate the projected matrix given
factors.
}
\seealso{
\code{\link{greg}}, \code{\link{roc}}, \code{\link{pgraph}}
}

