% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Compute the Projected Graph}
\usage{
roc(a, a0)
}
\arguments{
\item{a}{p * p estimated graph}

\item{a0}{p * p true graph}
}
\value{
a list.
\item{tpr}{tpr sequence}
\item{fpr}{fpr sequence}
}
\description{
\code{roc} calculate the fpr and tpr for the roc curve
}

